requirements.txt
flask
python-dotenv
pypdf==4.0.0
PyPDF2==3.0.1
pandas==2.2.0
chromadb==0.4.22
google-generativeai==0.3.2